feat: refactor issue management to use provider pattern
- Replaced direct GitHub CLI calls with a provider pattern for issue management, enhancing modularity and maintainability.
- Implemented methods for fetching, updating, commenting on, and creating issues in both GitHub and GitLab providers.
- Removed redundant TypeScript interface definitions and streamlined error handling across issue-related functions.
- Improved user feedback during issue operations by utilizing the provider's methods.
